+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-7130] Add include_links to concept search
+- [TD-7155] Introduce simple query string query for concepts
 
 ## [7.2.0] 2025-02-25